requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.0

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.0

# Validation
pydantic==2.9.0
pydantic-settings==2.5.0
email-validator==2.2.0

# Cache
redis==5.1.0
aioredis==2.0.1

# AI
openai==1.51.0
langsmith==0.1.125

# PDF generation
reportlab==4.2.2

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
httpx==0.27.0
